+++ packages/expo-video/CHANGELOG.md
@@ -61,6 +61,8 @@
 
 - [iOS] Support Apple TV. ([#29560](https://github.com/expo/expo/pull/29560) by [@douglowder](https://github.com/douglowder))
 
+- [iOS] SharedRef: migrate from `pointer` to `ref`
+
 ## 1.2.0 - 2024-06-20
 
 ### 🎉 New features